// frontend/src/api_client/models/UserRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { UserRole } from './UserRole';
export type UserRead = {
    id: string;
    email: string;
    is_active: boolean;
    role: UserRole;
    tenant_id: string;
};